requirements.txt
fastapi~=0.75.2
uvicorn~=0.17.6
sqlalchemy~=1.4.36
alembic~=1.7.7
psycopg2~=2.9.3
pyyaml~=6.0
argon2-cffi~=21.3.0
mdc~=1.2.1
starlette==0.17.1
pydantic~=1.9.0
jwcrypto~=1.3
email-validator~=1.2.1
Jinja2~=3.1.2
fastapi-redis-cache~=0.2.5
slowapi~=0.1.5